+++ spec/fred.clj
@@ -13,19 +13,23 @@
      [utils :as utils]]))
 
 (defn index
-  [^:injection loopback-handler partner]
+  [^:injection loopback-handler
+   partner]
   (go
     (-> (c/new-request loopback-handler)
         (c/to :get partner)
         c/send
         (c/then :pass-failure
-                :success [response (utils/response {:message (format "%s says `%s'" partner (-> response :body :message))})]))))
+                :success ([response] (utils/response {:message (format "%s says `%s'" partner (-> response :body :message))}))))))
 
-(defn show [id ^:injection loopback-handler partner]
+(defn show
+  [id
+   ^:injection loopback-handler
+   partner]
   (go
     (->
       (c/new-request loopback-handler)
       (c/to :get partner 123)
       c/send
       (c/then :pass-failure
-              :success [response (utils/response {:message (format "%s says `%s'" partner (-> response :body :message))})]))))
+              :success ([response] (utils/response {:message (format "%s says `%s'" partner (-> response :body :message))}))))))